fix: column spread prop type
When calculating the width for a column using the `spread` prop, the
value is divided by 12 and set to the percentage of the result:

https://github.com/jxom/bumbag-ui/blob/main/packages/bumbag/src/Columns/Columns.styles.ts#L62

The type for the prop only allows values 1-12, but float values are also
applicable for the prop. This PR generalises the type to allow any
number to be passed.